Blockchain&Backend/Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
web3
python-multipart
motor
websockets
cryptography
scikit-learn
python-jose